+++ Trelnex.Core.Data.Tests/CommandProviders/SqlCommandProviderTests.cs
@@ -7,14 +7,14 @@ using Snapshooter.NUnit;
 
 namespace Trelnex.Core.Data.Tests.CommandProviders;
 
-[Ignore("Requires a SQL instance.")]
+[Ignore("Requires a SQL server.")]
 public class SqlCommandProviderTests
 {
-    private readonly TokenCredential _tokenCredential = new DefaultAzureCredential();
     private readonly string _scope = "https://database.windows.net/.default";
-    private readonly string _typeName = "test-item";
 
+    private TokenCredential _tokenCredential = null!;
     private string _connectionString = null!;
+    private string _tableName = null!;
     private ICommandProvider<ITestItem> _commandProvider = null!;
 
     [OneTimeSetUp]
@@ -28,33 +28,34 @@ public class SqlCommandProviderTests
             .AddJsonFile("appsettings.User.json", optional: true, reloadOnChange: true)
             .Build();
 
-        var sqlConfiguration = configuration.GetSection("SQL").Get<SqlConfiguration>()!;
+        var providerConfiguration = configuration.GetSection("SqlCommandProviders").Get<SqlCommandProviderConfiguration>()!;
 
         var scsBuilder = new SqlConnectionStringBuilder()
         {
-            DataSource = sqlConfiguration.DataSource,
-            InitialCatalog = sqlConfiguration.InitialCatalog,
+            DataSource = providerConfiguration.DataSource,
+            InitialCatalog = providerConfiguration.InitialCatalog,
             Encrypt = true,
         };
 
         _connectionString = scsBuilder.ConnectionString;
+        _tableName = providerConfiguration.TableName;
 
         // create the command provider
-        var tokenCredential = new DefaultAzureCredential();
+        _tokenCredential = new DefaultAzureCredential();
 
         var sqlClientOptions = new SqlClientOptions(
-            TokenCredential: tokenCredential,
+            TokenCredential: _tokenCredential,
             Scope: _scope,
-            DataSource: sqlConfiguration.DataSource,
-            InitialCatalog: sqlConfiguration.InitialCatalog
+            DataSource: providerConfiguration.DataSource,
+            InitialCatalog: providerConfiguration.InitialCatalog
         );
 
         var factory = await SqlCommandProviderFactory.Create(
             sqlClientOptions);
 
         _commandProvider = factory.Create<ITestItem, TestItem>(
-            sqlConfiguration.TableName,
-            _typeName,
+            providerConfiguration.TableName,
+            "test-item",
             TestItem.Validator,
             CommandOperations.All);
     }
@@ -70,11 +71,11 @@ public class SqlCommandProviderTests
 
         sqlConnection.Open();
 
-        var sqlCommand = new SqlCommand("DELETE FROM [dbo].[test-items-events]; DELETE FROM [dbo].[test-items];", sqlConnection);
+        var cmdText = $"DELETE FROM [{_tableName}-events]; DELETE FROM [{_tableName}];";
+        var sqlCommand = new SqlCommand(cmdText, sqlConnection);
 
         sqlCommand.ExecuteNonQuery();
     }
-
 
     [Test]
     public async Task CreateCommand_SaveAsync()
@@ -848,7 +849,7 @@ public class SqlCommandProviderTests
     /// <summary>
     /// Represents the configuration properties for SQL command providers.
     /// </summary>
-    private record SqlConfiguration(
+    private record SqlCommandProviderConfiguration(
         string DataSource,
         string InitialCatalog,
         string TableName);